requirements.txt
asgiref==3.7.2
attrs==23.2.0
black==24.2.0
certifi==2024.2.2
click==8.1.7
colorama==0.4.6
distlib==0.3.8
Django==5.0.2
django-cors-headers==4.3.1
djangorestframework==3.14.0
filelock==3.13.1
flake8==7.0.0
install==1.3.5
interrogate==1.5.0
isort==5.13.2
mccabe==0.7.0
mypy-extensions==1.0.0
mysql-connector-python==8.3.0
packaging==23.2
pathspec==0.12.1
platformdirs==4.2.0
py==1.11.0
pycodestyle==2.11.1
pyflakes==3.2.0
PyMySQL==1.1.0
python-dotenv==1.0.1
pytz==2024.1
reverse==0.1.0
sqlparse==0.4.4
tabulate==0.9.0
toml==0.10.2
virtualenv==20.25.1
whitenoise==6.6.0
gunicorn